requirements.txt
farm-haystack
weaviate-client
fastapi
uvicorn
python-dotenv
sentence-transformers
llama-cpp-python
cached_property
farm-haystack[weaviate]
pypdf
torch
sentence_transformers
nltk
farm-haystack[preprocessing]
